Refactor, move to separate file.
